+++ test/stepper/stepper_widget_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide StepState, StepperType;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutterish/flutterish.dart';